[vuihandler]: Add a TX circular buffer to the BE. Fix an issue in the Injector
The main purpose of this fix is to have a way to queue the TX packets
before sending them in the vuhandler BE. I tis needed as we have
multiple sending sources, which reside both in the ME and the agency. It
was tested with an ME injection from the tablet, but still need to be
tested for the TX way.