[injector]: Remove the userspace ping-pong when injecting MEs from the kernel.
Following the previous commits, it removes and cleans everything which
was related to the kernel->userpsace->kernel injection when injecting
from the tablet (via BT).
The injection and is now fully done in the kernel Injector module. It
allocates a buffer the size of the ME at the start, receive it fully by
chunks (from the vuihandler) and the calls AVZ to inject it. This remove
the need to have an agency_read to read from the userspace and greatly
simplify the kernel Injector code.
It also remove everything not needed anymore in the userspace, such as
the IOCTLs numbers and the retreive_ME thread.