[soolink-debug]: Add netif_running and netif_oper_up check in the wlan plugin.
It should enforce the checks and avoid the wlan plugin to be stuck in a
infinite loop when the device is not operational.
All this is done in order to fix the bug which occurs on some RPi4, but
is still useful (and enforce the briing up check) so we should keep it.